test: ClientRequestCreator 분리
