+++ src/test/java/com/map/gaja/client/presentation/api/ClientControllerTest.java
@@ -110,7 +110,5 @@ class ClientControllerTest {
         Assertions.assertThat(validationErrorInfo.getCode()).isEqualTo("Size");
     }
 
-    @Test
-    void updateClient() {
-    }
+
 }
